Last but not the least : JF Fred2 ; correction de bugs, amelioration du code (standardisation)

+++ Ift232Tests/RandomGenTest.cs
@@ -7,7 +7,6 @@ namespace Ift232Tests
     [TestClass]
     public class RandomGenTest
     {
-
         [TestMethod]
         public void TestGetInstance()
         {
@@ -20,7 +19,7 @@ namespace Ift232Tests
         {
             RandomGen.SetToPredictable(5);
             int value = RandomGen.GetInstance().Next();
-            Assert.AreEqual(value,5);
+            Assert.AreEqual(value, 5);
         }
 
         [TestMethod]
@@ -36,7 +35,7 @@ namespace Ift232Tests
         {
             RandomGen.SetToPredictable(5);
             double value = RandomGen.GetInstance().NextDouble();
-            Assert.AreEqual(value, (double)5);
+            Assert.AreEqual(value, 5);
         }
     }
 }
